+++ Sources/Plasma/PubUtilLib/plDrawable/plSpaceTree.cpp
@@ -76,7 +76,7 @@ void plSpaceTreeNode::Write(hsStream* s)
     s->WriteLE16(fChildren[0]);
     if( fFlags & kIsLeaf )
         // Temp for now to play nice with binary patches
-        s->WriteLE16( 0 );
+        s->WriteLE16(uint16_t(0));
     else
         s->WriteLE16(fChildren[1]);
 }
@@ -534,7 +534,7 @@ void plSpaceTree::Read(hsStream* s, hsResMgr* mgr)
 
     fRoot = s->ReadLE16();
 
-    fNumLeaves = uint16_t(s->ReadLE32());
+    fNumLeaves = s->ReadLE32();
 
     uint32_t n = s->ReadLE32();
     fTree.SetCount(n);
